new js file for dom related functions also moved endCall function into webRTC functions file
